Fix std::pow ambiguity by casting both arguments on PS3
Casting both base and exponent to double and explicitly using std::pow to resolve ambiguity for the SNC compiler.

+++ RSDKv4/Script.cpp
@@ -2538,7 +2538,7 @@ void AppendIntegerToString(char *text, int value)
 
     v = 0;
     for (int i = cnt - 1; i >= 0; --i) {
-        v = value / (int)pow(10.0, i);
+        v = value / (int)std::pow(10.0, (double)i);
         v %= 10;
 
         int strValue = v + '0';
